# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smarketing CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: businessType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ═══════════════════════════════════
  # CONTACTS
  # ═══════════════════════════════════
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ═══════════════════════════════════
  # LEADS
  # ═══════════════════════════════════
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities
      parameters:
        - name: entityType
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: string
        - name: activityType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{id}:
    get:
      operationId: getActivity
      tags: [activities]
      summary: Get activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActivityBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ═══════════════════════════════════
  # NOTES
  # ═══════════════════════════════════
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: platform
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ═══════════════════════════════════
  # WORKFLOWS
  # ═══════════════════════════════════
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]

    # ─── USER ───────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        team:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        language:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, email, role, language, isActive, createdAt, updatedAt]
...
    Company:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        businessType:
          type: ["string", "null"]
        companySize:
          type: ["string", "null"]
        branchCount:
          type: ["integer", "null"]
        taxProfileStatus:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: string
        website:
          type: ["string", "null"]
        currentSystem:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, companyName, country, createdAt, updatedAt]
    CreateCompanyBody:
      type: object
      properties:
        companyName:
          type: string
        businessType:
          type: string
        companySize:
          type: string
        branchCount:
          type: integer
        taxProfileStatus:
          type: string
        city:
          type: string
        country:
          type: string
        website:
          type: string
        currentSystem:
          type: string
      required: [companyName]
    UpdateCompanyBody:
      type: object
      properties:
        companyName:
          type: string
        businessType:
          type: string
        companySize:
          type: string
        branchCount:
          type: integer
        taxProfileStatus:
          type: string
        city:
          type: string
        country:
          type: string
        website:
          type: string
        currentSystem:
          type: string

    # ─── CONTACT ────────────────────────
    Contact:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: ["string", "null"]
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        preferredContactMethod:
          type: string
        language:
          type: string
        isPrimary:
          type: boolean
        createdAt:
          type: string
      required: [id, fullName, preferredContactMethod, language, isPrimary, createdAt]
    CreateContactBody:
      type: object
      properties:
        companyId:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        jobTitle:
          type: string
        preferredContactMethod:
          type: string
        language:
          type: string
        isPrimary:
          type: boolean
      required: [fullName]
    UpdateContactBody:
      type: object
      properties:
        companyId:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        jobTitle:
          type: string
        preferredContactMethod:
          type: string
        language:
          type: string
        isPrimary:
          type: boolean

    # ─── LEAD ───────────────────────────
...
    Activity:
      type: object
      properties:
        id:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        userId:
          type: ["string", "null"]
        activityType:
          type: string
        direction:
          type: ["string", "null"]
        activityDatetime:
          type: string
        subject:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        interestLevel:
          type: ["string", "null"]
        aiSummary:
          type: ["string", "null"]
        aiSentiment:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, entityType, entityId, activityType, activityDatetime, createdAt]
    CreateActivityBody:
      type: object
      properties:
        entityType:
          type: string
        entityId:
          type: string
        userId:
          type: string
        activityType:
          type: string
        direction:
          type: string
        activityDatetime:
          type: string
        subject:
          type: string
        content:
          type: string
        outcome:
          type: string
        nextStep:
          type: string
        interestLevel:
          type: string
      required: [entityType, entityId, activityType]
    UpdateActivityBody:
      type: object
      properties:
        subject:
          type: string
        content:
          type: string
        outcome:
          type: string
        nextStep:
          type: string
        interestLevel:
          type: string
        aiSummary:
          type: string
        aiSentiment:
          type: string

    # ─── NOTE ────────────────────────────
...
    Campaign:
      type: object
      properties:
        id:
          type: string
        campaignKey:
          type: ["string", "null"]
        platform:
          type: ["string", "null"]
        campaignName:
          type: string
        campaignObjective:
          type: ["string", "null"]
        status:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, campaignName, status, createdAt]
    CreateCampaignBody:
      type: object
      properties:
        campaignKey:
          type: string
        platform:
          type: string
        campaignName:
          type: string
        campaignObjective:
          type: string
        status:
          type: string
        startDate:
          type: string
        endDate:
          type: string
      required: [campaignName]
    UpdateCampaignBody:
      type: object
      properties:
        campaignKey:
          type: string
        platform:
          type: string
        campaignName:
          type: string
        campaignObjective:
          type: string
        status:
          type: string
        startDate:
          type: string
        endDate:
          type: string

    # ─── WORKFLOW ────────────────────────
...